♻️ [Refactoring]: #17 IndirectRef を pdf/types/indirect-ref/ に集約し単一ソース化
これまで型定義は pdf-types/index.ts、companion object は page-tree/indirect-ref.ts に
分散していたため「同じドメイン内に IndirectRef が 2 つ存在」する状態だった。
ObjectNumber / GenerationNumber と同じパターンで pdf/types/indirect-ref/index.ts に
型 + companion object を同居させ、唯一の真実とする。pdf-types/index.ts は型のみ
re-export して既存 consumer の import パスを維持する。
