+++ packages/core/src/document/page-tree/dict-reader.ts
@@ -1,7 +1,7 @@
 import { NumberEx } from "../../ext/number/index";
+import { IndirectRef } from "../../pdf/types/indirect-ref/index";
 import type { PdfObject, PdfValue } from "../../pdf/types/pdf-types/index";
 import { none, type Option, some } from "../../utils/option/index";
-import { IndirectRef } from "./indirect-ref";
 import type { PdfRectangle } from "./resolved-page";
 
 const BOX_ELEMENT_COUNT = 4;
