+++ packages/core/src/pdf/types/index.ts
@@ -1,4 +1,5 @@
 export { ByteOffset } from "./byte-offset/index";
 export { GenerationNumber } from "./generation-number/index";
+export { IndirectRef } from "./indirect-ref/index";
 export { ObjectNumber } from "./object-number/index";
 export * from "./pdf-types/index";
